customer-python-cert/requirements.txt
azure-identity>=1.23,<2.0
azure-keyvault-secrets>=4.10,<5.0
msal>=1.31,<2.0
python-dotenv>=1.1,<2.0
requests>=2.32,<3.0
